Dragon Egg fix and improvements
The dragon egg now longer spawns on top of the 1st one generated by vanilla. Now the game just gives the egg to the player that killed the dragon. If the admin score isn't set to at least 2, the status of killing the dragon doesn't reset, which means they'll only get the egg once.