Real Pet Chatter Overhauled for performance + Silence Mob Trigger Commands
The player now sets pets to be a "chatting mob". It using a "chatting_mob_v2" tag, to for those who are updating, due to the fact that the previous version would of already tagged many wild mobs. The chatting mobs will run the functions, instead of each function testing if the mob is a chatting mob.

To help with large noisy mob farms that don't have controlled chatting, the silence mob pack now has more possibilities.